feat : nonchain 추상화
