+++ src/main/java/nextstep/auth/AuthConfig.java
@@ -1,12 +1,12 @@
 package nextstep.auth;
 
-import nextstep.auth.authentication.BasicAuthenticationFilter;
-import nextstep.auth.authentication.BearerTokenAuthenticationFilter;
+import nextstep.auth.interceptor.BasicAuthenticationFilter;
+import nextstep.auth.interceptor.BearerTokenAuthenticationFilter;
 import nextstep.auth.authentication.UsernamePasswordAuthenticationFilter;
 import nextstep.auth.authorization.AuthenticationPrincipalArgumentResolver;
 import nextstep.auth.context.SecurityContextPersistenceFilter;
 import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenAuthenticationInterceptor;
+import nextstep.auth.interceptor.TokenAuthenticationInterceptor;
 import nextstep.member.application.LoginMemberService;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
