+++ src/main/java/nextstep/auth/interceptor/AuthenticationChainHandler.java
@@ -1,0 +1,25 @@
+package nextstep.auth.interceptor;
+
+import nextstep.auth.context.Authentication;
+import nextstep.auth.context.SecurityContextHolder;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public abstract class AuthenticationChainHandler implements HandlerInterceptor {
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+        try {
+            System.out.println("chain handler preHandle");
+            Authentication authentication = createAuthentication(request);
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+            return true;
+        } catch (Exception e) {
+            return true;
+        }
+    }
+
+    protected abstract Authentication createAuthentication(HttpServletRequest request);
+}
