+++ src/main/java/nextstep/auth/interceptor/BasicAuthenticationFilter.java
@@ -1,0 +1,44 @@
+package nextstep.auth.interceptor;
+
+import nextstep.auth.authentication.AuthenticationToken;
+import nextstep.auth.authentication.AuthorizationExtractor;
+import nextstep.auth.authentication.AuthorizationType;
+import nextstep.auth.context.Authentication;
+import nextstep.auth.exception.AuthenticationException;
+import nextstep.member.application.LoginMemberService;
+import nextstep.member.domain.LoginMember;
+import org.apache.tomcat.util.codec.binary.Base64;
+
+import javax.servlet.http.HttpServletRequest;
+
+public class BasicAuthenticationFilter extends AuthenticationChainHandler {
+    private final LoginMemberService loginMemberService;
+
+    public BasicAuthenticationFilter(LoginMemberService loginMemberService) {
+        this.loginMemberService = loginMemberService;
+    }
+
+    @Override
+    protected Authentication createAuthentication(HttpServletRequest request) {
+
+        String authCredentials = AuthorizationExtractor.extract(request, AuthorizationType.BASIC);
+        String authHeader = new String(Base64.decodeBase64(authCredentials));
+
+        String[] splits = authHeader.split(":");
+        String principal = splits[0];
+        String credentials = splits[1];
+
+        AuthenticationToken token = new AuthenticationToken(principal, credentials);
+
+        LoginMember loginMember = loginMemberService.loadUserByUsername(token.getPrincipal());
+        if (loginMember == null) {
+            throw new AuthenticationException();
+        }
+
+        if (!loginMember.checkPassword(token.getCredentials())) {
+            throw new AuthenticationException();
+        }
+
+        return new Authentication(loginMember.getEmail(), loginMember.getAuthorities());
+    }
+}
