+++ src/test/java/nextstep/subway/acceptance/AuthAcceptanceTest.java
@@ -4,7 +4,6 @@ import io.restassured.RestAssured;
 import io.restassured.authentication.FormAuthConfig;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
@@ -37,7 +36,6 @@ class AuthAcceptanceTest extends AcceptanceTest {
     @DisplayName("Bearer Auth")
     @Test
     void myInfoWithBearerAuth() {
-//        String accessToken = 로그인_되어_있음(ADMIN_EMAIL, ADMIN_PASSWORD);
 
         ExtractableResponse<Response> response = 베어러_인증으로_내_회원_정보_조회_요청(관리자토큰);
 
@@ -47,12 +45,13 @@ class AuthAcceptanceTest extends AcceptanceTest {
     @DisplayName("유효하지 않은 토큰으로 인한 Bearer Auth 실패.")
     @Test
     void myInfoWithBearerAuthFail() {
-//        String accessToken = 로그인_되어_있음(ADMIN_EMAIL, ADMIN_PASSWORD);
         String invalidToken = "invalidToken";
+
         ExtractableResponse<Response> response = 베어러_인증으로_내_회원_정보_조회_요청(invalidToken);
 
         assertThat(response.statusCode()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
     }
+
 
     private ExtractableResponse<Response> 폼_로그인_후_내_회원_정보_조회_요청(String email, String password) {
         return RestAssured.given().log().all()
